+++ activerecord/lib/active_record/connection_adapters/postgresql/database_statements.rb
@@ -175,13 +175,14 @@ module ActiveRecord
               return ActiveRecord::Result.empty
             end
 
-            types = {}
             fields = result.fields
-            fields.each_with_index do |fname, i|
-              ftype = result.ftype i
-              fmod  = result.fmod i
-              types[fname] = types[i] = get_oid_type(ftype, fmod, fname)
+            types = Array.new(fields.size)
+            fields.size.times do |index|
+              ftype = result.ftype(index)
+              fmod  = result.fmod(index)
+              types[index] = get_oid_type(ftype, fmod, fields[index])
             end
+
             ar_result = ActiveRecord::Result.new(fields, result.values, types.freeze)
             result.clear
             ar_result